+++ RR-Common/src/main/java/com/scor/rr/repository/RranalysisRepository.java
@@ -3,7 +3,7 @@ package com.scor.rr.repository;
 import com.scor.rr.domain.RrAnalysisEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RranalysisRepository extends JpaRepository<RrAnalysisEntity, Integer> {
-    RrAnalysisEntity findByAnalysisId(Integer s);
+public interface RranalysisRepository extends JpaRepository<RrAnalysisEntity, Long> {
+    RrAnalysisEntity findByAnalysisId(Long s);
 }
